requirements.txt
torch>=1.10
torchvision
numpy
PyYAML
easydict
opencv-python
Pillow
scikit-learn
scikit-image
huggingface_hub
einops
kornia
more_itertools
gradio==3.50.2
loralib
tabulate
safetensors
transformers==4.56.0
timm==1.0.20
accelerate==1.10.1
git+https://github.com/SunzeY/AlphaCLIP